Problem: handshake killed 3rd innocent node
When a node received a "HAVE WRITER" and then "HAVE READER" from
the cluster, it would falsely try to become a pipe writer. This
caused an assertion failure on zpipes_server.c:262 as the code
caught this impossible situation.

Solution: when seeing two peers handshake the same pipe, drop the
pipe entirely.

Fixes #82.

Note: I didn't set up a reproducible test case so I'm not 100%
sure the fix is accurate.